[#2037] Added PgSQL migration 7
This migration adds new columns reclaimed to the lease4 and lease6 tables
that reflects whether the lease is in the expired-reclaimed state or not.
Having such column allows for creating a composite index using this column
and the expiration time. The old indexes by state and expire time were
removed.